refactor: per-test Autentico isolation for deterministic E2E tests
Move Autentico lifecycle (onboard, start, seed, stop) from global
setup into a Playwright fixture that runs for each test. Every test
gets a pristine IdP with a fresh database — no session accumulation
or state leakage between tests. This ensures any test failure is a
real bug in the adapter, not server-side noise.

Global setup now only downloads the binary and runs init (both cached).
Added workers: 1 since all tests share port 9999.

Overhead: ~375ms per test (22s vs 16s for 16 tests).
