+++ tests/e2e/playwright.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   testDir: "./specs",
   timeout: 30_000,
   retries: 0,
+  workers: 1,
   reporter: process.env.CI ? [["list"], ["html", { open: "never" }]] : "list",
   use: {
     baseURL: "http://localhost:5173",
